fix: keep caption shortcuts outside component
